+++ src/app/stats/popular-commands-graph/popular-commands-graph.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { PopularCommandsGraphComponent } from './popular-commands-graph.component';
+
+describe('PopularCommandsGraphComponent', () => {
+  let component: PopularCommandsGraphComponent;
+  let fixture: ComponentFixture<PopularCommandsGraphComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ PopularCommandsGraphComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(PopularCommandsGraphComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});